Add Review Feature at the Domain layer add repositories contracts needed Fix rating data type

+++ HotelBookingSystem.Domain/Models/Review.cs
@@ -9,6 +9,5 @@ public class Review : Entity
     public Guid HotelId { get; set; }
     public string? Title { get; set; }
     public string Description { get; set; } = default!;
-
-    public short Rating { get; set; } 
+    public int Rating { get; set; } 
 }
